added highlight script, make drawers and lever ray sensible and print click on highlight objects
