+++ altlas/Assets/Prefabs/Map_alt.prefab
@@ -22,8 +22,9 @@ GameObject:
   - component: {fileID: 33419303764663426}
   - component: {fileID: 23282047421209194}
   - component: {fileID: 65802085699699740}
-  - component: {fileID: 54498630221748466}
   - component: {fileID: 114163787465769024}
+  - component: {fileID: 54611382421704714}
+  - component: {fileID: 114381881185707782}
   m_Layer: 0
   m_Name: Map_alt
   m_TagString: Pickupable
@@ -88,7 +89,7 @@ MeshFilter:
   m_Mesh: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
 --- !u!54 &54611382421704714
 Rigidbody:
-  m_ObjectHideFlags: 1
+  m_ObjectHideFlags: 0
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1141682715052752}
@@ -122,6 +123,31 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: c32a57a6a8c877f43ba1d8bf79b314b1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  data:
+    m_year: 0
+    m_language: 
+    m_coordinate: 
+    m_title: 
+    m_source: 
+    m_imageSize: 
+    m_property: 
+    m_description: 
+    m_category: 
+    m_subCategory: 
+    m_location: 
+    m_id: 
+    texture: {fileID: 0}
+--- !u!114 &114381881185707782
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1141682715052752}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a010a2245178d45a3bce04e6a7eb0be1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   rend: {fileID: 0}
