+++ altlas/Assets/scripts/HighlightScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a010a2245178d45a3bce04e6a7eb0be1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
